requirements.txt
biopython
scikit-learn

# pubmed baseline download
requests

# neural retrieval
transformers
torch>=2.4 

#pubmed
biopython


# elastic search
elasticsearch
lxml
dotenv

# statistics
prettytable

# visualization
matplotlib

# synonym mapping 
nltk

# evaluation
pytrec_eval

#optuna
optuna
plotly

joblib
scipy

datasets
jsonpickle
numba
retriv
ranx

sentence-transformers
torchvision

numpy

# Utilities
tqdm
regex

pandas
scikit-learn
jsonpickle